get refactor to a point of passing most tests and remove all migrations
