+++ baleen/job/admin.py
@@ -1,7 +1,7 @@
 from django.contrib import admin
 
 from baleen.job.models import Job
-from baleen.action.models import ActionResult
+from baleen.project.models import ActionResult
 
 class ActionResultInline(admin.ModelAdmin):
     model = ActionResult
